EditAccount: ugly patch (originally written in 2013) to store information about disabled accounts (namely the 'disabled' and 'disabled_date' preferences) into the global_preferences table (in addition to ShoutWiki Hub's local prefs table, user_properties).
This way the "This account has been disabled by ShoutWiki staff" notice will show up on *all* wikis instead of just Hub, and maybe, just maybe local admins won't bother blocking disabled accounts as eagerly as they do now.

setDisabledFlag.php needs to be run with the --doit option to set the 'disabled' and 'disabled_date' global preferences for accounts where the real name has been set to "Account Disabled". I admit it's not the most robust way of doing this, but this script is a one-shot and it's rather safe to assume that humans (or spambots!) have no real reason to set their real name to "Account Disabled".